// dashboard/src/types.ts
export type URLTab = {
  url: string;
  title: string;
  icon: LucideIcon | TablerIcon;
};